+++ idea/src/org/jetbrains/kotlin/idea/refactoring/safeDelete/SafeDeleteImportDirectiveUsageInfo.kt
@@ -16,17 +16,27 @@
 
 package org.jetbrains.kotlin.idea.refactoring.safeDelete
 
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiMember
 import com.intellij.refactoring.safeDelete.usageInfo.SafeDeleteReferenceSimpleDeleteUsageInfo
+import org.jetbrains.kotlin.asJava.unwrapped
+import org.jetbrains.kotlin.idea.caches.resolve.getJavaMemberDescriptor
 import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptor
 import org.jetbrains.kotlin.idea.core.targetDescriptors
 import org.jetbrains.kotlin.psi.KtDeclaration
 import org.jetbrains.kotlin.psi.KtImportDirective
 
 class SafeDeleteImportDirectiveUsageInfo(
-        importDirective: KtImportDirective, declaration: KtDeclaration
+        importDirective: KtImportDirective, declaration: PsiElement
 ) : SafeDeleteReferenceSimpleDeleteUsageInfo(importDirective, declaration, importDirective.isSafeToDelete(declaration))
 
-private fun KtImportDirective.isSafeToDelete(declaration: KtDeclaration): Boolean {
+private fun KtImportDirective.isSafeToDelete(element: PsiElement): Boolean {
     val referencedDescriptor = targetDescriptors().singleOrNull() ?: return false
-    return referencedDescriptor == declaration.resolveToDescriptor()
+    val unwrappedElement = element.unwrapped
+    val declarationDescriptor = when (unwrappedElement) {
+        is KtDeclaration -> unwrappedElement.resolveToDescriptor()
+        is PsiMember -> unwrappedElement.getJavaMemberDescriptor()
+        else -> return false
+    }
+    return referencedDescriptor == declarationDescriptor
 }