+++ ChangeLog.md
@@ -86,6 +86,7 @@
 - [`KT-14854`](https://youtrack.jetbrains.com/issue/KT-14854) Extract Interface: Fix NPE on dialog opening
 - [`KT-14814`](https://youtrack.jetbrains.com/issue/KT-14814) Rename: Fix renaming of .kts file to .kt and vice versa
 - [`KT-14361`](https://youtrack.jetbrains.com/issue/KT-14361) Rename: Do not report redeclaration conflict for private top-level declarations located in different files
+- [`KT-14596`](https://youtrack.jetbrains.com/issue/KT-14596) Safe Delete: Fix exception on deleting Java class used in Kotlin import directive(s)
 
 #### J2K
 - [`KT-14604`](https://youtrack.jetbrains.com/issue/KT-14604) Floating-point literals: "dot" + "exponent" char sequence was processed incorrectly